Split main policy sections (Inbound, outbound...) into partial classes by policy
