+++ src/Oriflame.PolicyBuilder.Xml/Builders/Sections/BackendSectionPolicyBuilder.Choose.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using Oriflame.PolicyBuilder.Policies.Builders.Fluent.Sections;
+using Oriflame.PolicyBuilder.Policies.Definitions;
+
+namespace Oriflame.PolicyBuilder.Xml.Builders.Sections
+{
+    public partial class BackendSectionPolicyBuilder : SectionBuilderBase<IBackendSectionPolicyBuilder>, IBackendSectionPolicyBuilder
+    {
+        /// <inheritdoc />
+        public override IBackendSectionPolicyBuilder Choose(Func<IConditionSectionBuilder<IBackendSectionPolicyBuilder>, ISectionPolicy> conditionBuilder)
+        {
+            var conditionSectionBuilder = new BackendConditionSectionBuilder();
+            return AddPolicyDefinition(conditionBuilder.Invoke(conditionSectionBuilder));
+        }
+    }
+}
