+++ src/Oriflame.PolicyBuilder.Xml/Builders/Sections/BackendSectionPolicyBuilder.SetBody.cs
@@ -1,12 +1,15 @@
 ﻿using Oriflame.PolicyBuilder.Policies.Builders.Fluent.Sections;
-using Oriflame.PolicyBuilder.Policies.Definitions;
+using Oriflame.PolicyBuilder.Policies.DynamicProperties;
+using Oriflame.PolicyBuilder.Xml.Definitions.Common;
 
 namespace Oriflame.PolicyBuilder.Xml.Builders.Sections
 {
     public partial class BackendSectionPolicyBuilder : SectionBuilderBase<IBackendSectionPolicyBuilder>, IBackendSectionPolicyBuilder
     {
-        public BackendSectionPolicyBuilder(ISectionPolicy sectionPolicy) : base(sectionPolicy)
+        /// <inheritdoc />
+        public virtual IBackendSectionPolicyBuilder SetBody(ILiquidTemplate template)
         {
+            return AddPolicyDefinition(new SetBodyLiquid(template));
         }
     }
 }
